move search from user into domain
