+++ app/src/main/java/com/javahelp/frontend/domain/search/ISearchOutput.java
@@ -1,4 +1,4 @@
-package com.javahelp.frontend.domain.user.search;
+package com.javahelp.frontend.domain.search;
 
 
 import com.javahelp.model.survey.SurveyResponse;
