+++ app/src/main/java/com/javahelp/frontend/view/SearchFragment.java
@@ -17,7 +17,7 @@ import androidx.lifecycle.ViewModelProvider;
 import com.google.android.material.chip.Chip;
 import com.javahelp.R;
 import com.javahelp.frontend.domain.user.login.LoginResult;
-import com.javahelp.frontend.domain.user.search.SearchResult;
+import com.javahelp.frontend.domain.search.SearchResult;
 import com.javahelp.model.survey.Survey;
 import com.javahelp.model.survey.SurveyQuestion;
 import com.javahelp.model.survey.SurveyQuestionResponse;
